refactor: FrameStatus 를 제공하도록 변경
